refactor: CoursePageContent fix: swipe to dismiss bugs
