+++ app/src/main/java/com/dscvit/vitty/data/converter/Converters.kt
@@ -1,17 +1,12 @@
 package com.dscvit.vitty.data.converter
 
 import androidx.room.TypeConverter
-import com.dscvit.vitty.ui.coursepage.NoteType
+import com.dscvit.vitty.ui.coursepage.models.NoteType
 
 class Converters {
-    
     @TypeConverter
-    fun fromNoteType(noteType: NoteType): String {
-        return noteType.name
-    }
-    
+    fun fromNoteType(noteType: NoteType): String = noteType.name
+
     @TypeConverter
-    fun toNoteType(noteType: String): NoteType {
-        return NoteType.valueOf(noteType)
-    }
+    fun toNoteType(noteType: String): NoteType = NoteType.valueOf(noteType)
 }
